wrap author.img in a div to make it block
wrap author.image in a div to make it block style. The demo portion does a placekitten of 200x300, but since it's inlined, the spec'd res isn't shown, which may confuse some developers. Additionally, it leads to the pic bleeding down below the border of the element (which is easily seen either with devtools or by doing :host{ background-color: yellow; } in the seed-element CSS.
So I'm proposing moving to block style by wrapping it in a div. Other option would be to make it block style via explict CSS but then you'd need another class added in.